Moved getting location to dashboard activity
